Refactor styled components back into regular component files
