+++ src/client/components/NavBar.jsx
@@ -1,32 +1,78 @@
 import React from 'react';
-import styled from 'styled-components';
+import styled, { keyframes } from 'styled-components';
 import PropTypes from 'prop-types';
 
-import { NavBarContentSplitTwoSides } from '../styled-components/basicNavStyles';
-import { CenterVertically } from '../styled-components/layoutStyles';
+
+const RowLayout = styled.div`
+  display: flex;
+  padding: 15px 0;
+
+  > * {
+    margin: 0 15px;
+  }
+`;
+
+const CenterVertically = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+const CenteredPaddedChildren = styled.div`
+  display: flex;
+  align-items: center;
+
+  > * {
+    margin-left: 20px;
+  }
+`;
+
+const NavBarContentSplitTwoSides = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: ${props => props.theme.primaryColor};
+  color: ${props => props.theme.contrastColor};
+  padding: 0 30px;
+  height: 60px;
+`;
+
+const rotate = keyframes`
+  from {
+    transform: rotate(0deg);
+  }
+
+  to {
+    transform: rotate(360deg);
+  }
+`;
+
+const Rotate = styled.div`
+  animation: ${rotate} 10s linear infinite;
+`;
+
 
 const MainNavElement = styled.div`
   margin-right: 20px;
   width: 70px;
   text-align: center;
+
   cursor: pointer;
   :hover {
-    font-weight: bold;
+    color: #FCEDBF;
   }
-`;
 
-const LeftNavElement = styled.div`
-  margin-left: 20px;
+  transition-duration: 500ms;
+  transition-property: color;
 `;
 
 const NavBar = ({ subModuleTitle }) => (
   <NavBarContentSplitTwoSides>
-    <CenterVertically>
-      <img src="nav-gear-white.png" alt="logo" />
-      <LeftNavElement>The high art of getting appointments</LeftNavElement>
-      <LeftNavElement>{">"}</LeftNavElement>
-      <LeftNavElement>{subModuleTitle}</LeftNavElement>
-    </CenterVertically>
+    <CenteredPaddedChildren>
+      <Rotate><img src="nav-gear-white.png" alt="logo" /></Rotate>
+      <div>The high art of getting appointments</div>
+      <div>{'>'}</div>
+      <div>{subModuleTitle}</div>
+    </CenteredPaddedChildren>
     <CenterVertically>
       <MainNavElement>Modules</MainNavElement>
       <MainNavElement>Settings</MainNavElement>
@@ -40,3 +86,14 @@ NavBar.propTypes = {
 };
 
 export default NavBar;
+
+/*
+<CenterVertically>
+      <Rotate>
+        <img src="nav-gear-white.png" alt="logo" />
+      </Rotate>
+      <LeftNavElement>The high art of getting appointments</LeftNavElement>
+      <LeftNavElement>{'>'}</LeftNavElement>
+      <LeftNavElement>{
+
+*/
