--- src/client/styled-components/layoutStyles.jsx
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-const RowLayout = styled.div`
-  display: flex;
-  padding: 15px 0;
-
-  > * {
-    margin: 0 15px;
-  }
-`;
-
-const CenterVertically = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-export { RowLayout, CenterVertically };
